backend/requirements.txt
asgiref==3.8.1
astroid==3.1.0
cfgv==3.4.0
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.14.0
identify==2.5.36
importlib_metadata==7.1.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.1
pre-commit==3.7.0
pylint==3.1.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.1
yapf==0.40.2
zipp==3.18.1
